Refining Turn Manager Operations
Added a function for converting the active entitiy list to the TurnOrder. The function also sorts the Player to the top of the list. I think we'll need to create an execute turn function next, which may mean we need to figure out the specific details of turns relative to the player. (also cleaned up the TileManager a smidge)