Action Setup Tangent
This should really be its own branch, but I felt weird changing branches with so much of the broken Action Setup branch carried over. Anyway, I put together a simple TileInventory and functioinality for dragging those objects out to be placed on the grid. Still need to incorporate this ability into the action system to spend AP.